pyproject.toml
[project]
name = "newshound"
version = "0.1.0"
description = "Discord News Feed Bot"
authors = [
    { name = "Yusei Ito", email = "me@yuseiito.com" }
]
dependencies = [
    "discord.py",
    "feedparser",
    "python-dotenv",
    "alembic",
    "sqlalchemy"
]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "GPLv3" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-url = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.wheel.sources."src"]
type = "directory"
